% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eta1G.f}
\alias{eta1G.f}
\title{First Derivative Eta1 (Internal)}
\usage{
eta1G.f(epn, y, H1)
}
\arguments{
\item{epn}{Numeric value, e.g., \eqn{\exp(Z \beta)}.}

\item{y}{Numeric scalar or vector.}

\item{H1}{A function. The first derivative of \code{H}.}
}
\value{
A numeric value.
}
\description{
Computes \eqn{-epn \times H1(epn \times y)}, typically used in partial derivatives of log-likelihood.
}
\keyword{internal}
