% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eta2G.f}
\alias{eta2G.f}
\title{Second Derivative Eta2 (Internal)}
\usage{
eta2G.f(epn, y, H2)
}
\arguments{
\item{epn}{Numeric value, e.g., \eqn{\exp(Z \beta)}.}

\item{y}{Numeric scalar or vector.}

\item{H2}{A function. The second derivative of \code{H}.}
}
\value{
A numeric value.
}
\description{
Computes \eqn{-epn^2 \times H2(epn \times y)}.
}
\keyword{internal}
