% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{H.func}
\alias{H.func}
\title{Baseline Hazard Function Constructor (Internal)}
\usage{
H.func(G, G1, G2)
}
\arguments{
\item{G}{A function \code{G(x)}, e.g., a cumulative hazard transform.}

\item{G1}{A function, the first derivative of \code{G}.}

\item{G2}{A function, the second derivative of \code{G}.}
}
\value{
A list with three functions:
  \describe{
    \item{\code{H(x)}}{Computes \eqn{\exp(-G(x))}.}
    \item{\code{H1(x)}}{First derivative of \code{H(x)}.}
    \item{\code{H2(x)}}{Second derivative of \code{H(x)}.}
  }
}
\description{
Given functions \code{G}, \code{G1}, and \code{G2}, returns a list of baseline
functions \code{H}, \code{H1}, and \code{H2} used in survival or hazard calculations.
}
\keyword{internal}
