% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntCens-package.R
\docType{package}
\name{IntCens-package}
\alias{IntCens}
\alias{IntCens-package}
\title{IntCens: Non- and Semi-Parametric Analysis of Interval-Censored Data}
\description{
Implements non- and semi-parametric methods for analyzing interval-censored data. This includes iterative estimation routines for proportional hazards and proportional odds models, as well as nonparametric maximum likelihood estimation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lmaowisc/IntCens}
  \item \url{https://lmaowisc.github.io/IntCens/}
}

}
\keyword{internal}
