% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{plot.icsurvfit}
\alias{plot.icsurvfit}
\title{Plot Method for icsurvfit Objects}
\usage{
\method{plot}{icsurvfit}(
  object,
  z = NULL,
  xlab = "Time",
  ylab = "Survival rate",
  lty = 1,
  frame.plot = FALSE,
  add = FALSE,
  ylim = c(0, 1),
  ...
)
}
\arguments{
\item{object}{An object of class \code{"icsurvfit"} returned by
\code{\link{icsurvfit}}.}

\item{z}{Numeric (vector of) covariate values for which to plot the survival
curve. If \code{NULL} (default) and \code{model} is nonparametric,
plots the baseline curve.}

\item{xlab, ylab}{Axis labels.}

\item{lty}{Line type. Default is \code{1}.}

\item{frame.plot}{Logical indicating whether to draw a box around the plot.}

\item{add}{Logical indicating whether to add the curve to an existing plot.
If \code{FALSE} (default), a new plot is created.}

\item{ylim}{Numeric vector of length 2 specifying the \eqn{y}-axis range.}

\item{...}{Additional graphical parameters passed to \code{\link{plot}} or
\code{\link{lines}}.}
}
\description{
Plots the estimated survival function from an \code{icsurvfit} object.
If \code{model="NP"}, the nonparametric estimate is plotted. For
\code{model="PO"} or \code{"PH"}, the survival curve for a specific covariate
value \code{z} can be overlaid. By default, it plots a step function.
}
