% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.icsurvfit}
\alias{print.icsurvfit}
\title{Print Method for icsurvfit Objects}
\usage{
\method{print}{icsurvfit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"icsurvfit"} returned by
\code{\link{icsurvfit}}.}

\item{...}{Further arguments passed to or from other methods. Not currently used.}
}
\description{
Prints a summary of the fitted model, including the call, convergence status,
estimated survival function (for \code{model="NP"}), or estimated regression
coefficients (for \code{model="PO"} or \code{"PH"}).
}
