% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trans_inv}
\alias{trans_inv}
\title{Inverse Log-Transform Function Constructor (Internal)}
\usage{
trans_inv(gamma)
}
\arguments{
\item{gamma}{Numeric scalar. If 0, the inverse transform is the identity function.}
}
\value{
A single function, \code{Ginv}, which computes the inverse transform.
}
\description{
Creates an inverse function for the transformations constructed by \code{\link{trans_log}}.
}
\keyword{internal}
