% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trans_log}
\alias{trans_log}
\title{Log-Transform Functions Constructor (Internal)}
\usage{
trans_log(gamma)
}
\arguments{
\item{gamma}{Numeric scalar. If \code{gamma = 0}, the transform is identity; otherwise,
a log-based transform is used.}
}
\value{
A list of 5 functions:
  \describe{
    \item{\code{transform[[1]]}}{The main transform.}
    \item{\code{transform[[2]]}}{The first derivative of the transform.}
    \item{\code{transform[[3]]}}{The second derivative.}
    \item{\code{transform[[4]]}}{The third derivative.}
    \item{\code{transform[[5]]}}{The fourth derivative.}
  }
}
\description{
Creates a list of transformation functions and their derivatives based on the
parameter \code{gamma}. Used to transform (and invert) data in estimation routines.
}
\keyword{internal}
