% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{xi1G.f}
\alias{xi1G.f}
\title{First Derivative of Beta (xi1) (Internal)}
\usage{
xi1G.f(epn, z, y, H1)
}
\arguments{
\item{epn}{Numeric, e.g. \eqn{\exp(Z \beta)}.}

\item{z}{Numeric or vector, representing the covariate(s).}

\item{y}{Numeric, typically \eqn{y[j]} in iterative algorithms.}

\item{H1}{A function, the first derivative of \code{H}.}
}
\value{
A numeric value for \code{xi1}.
}
\description{
Computes \eqn{-z \times epn \times y \times H1(epn \times y)}, used in partial derivative w.r.t. \code{beta}.
}
\keyword{internal}
