% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{xi2G.f}
\alias{xi2G.f}
\title{Second Derivative of Beta (xi2) (Internal)}
\usage{
xi2G.f(epn, z, y, H1, H2)
}
\arguments{
\item{epn}{Numeric, e.g. \eqn{\exp(Z \beta)}.}

\item{z}{Numeric or vector, representing the covariate(s) (possibly multiple dimensions).}

\item{y}{Numeric, typically \eqn{y[j]}.}

\item{H1, H2}{Functions, first and second derivatives of \code{H}.}
}
\value{
A matrix or numeric value, the second derivative term.
}
\description{
Computes \eqn{-z z^T \times epn \times y \times (H1(epn \times y) + y \times epn \times H2(epn \times y))}.
}
\keyword{internal}
