% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{zeta21.f}
\alias{zeta21.f}
\title{Zeta21 Calculation (Internal)}
\usage{
zeta21.f(Fd1, Fd2, eta1, eta2)
}
\arguments{
\item{Fd1}{Numeric value (lower bound CDF).}

\item{Fd2}{Numeric value (upper bound CDF).}

\item{eta1}{Numeric, e.g., \code{eta1G.f(epn, x, H1)}.}

\item{eta2}{Numeric, e.g., \code{eta2G.f(epn, x, H2)}.}
}
\value{
A numeric value for the zeta21 term.
}
\description{
Computes \eqn{(Fd2 - Fd1)^{-2} \times \eta1^2 + (Fd2 - Fd1)^{-1} \times \eta2}.
}
\keyword{internal}
